working bme280

+++ Sensor/timer.c
@@ -5,7 +5,7 @@
  *      Author: Lukas
  */
 # include "timer.h"
-void delay_ms(volatile uint16 time){
+void delay_ms(volatile uint32 time){
 	time *= time;
 	while(time){
 		time--;
